refactor: remove requests from event forwarding
