+++ mergify_engine/exceptions.py
@@ -13,7 +13,6 @@
 # under the License.
 
 import httpx
-import requests
 
 
 class MergifyNotInstalled(Exception):
@@ -42,18 +41,8 @@ def need_retry(exception):  # pragma: no cover
         return max(exception.countdown, RATE_LIMIT_RETRY_MIN)
     elif isinstance(exception, MergeableStateUnknown):
         return BASE_RETRY_TIMEOUT
-    elif (
-        (
-            isinstance(exception, httpx.HTTPError)
-            and (exception.response is None or exception.response.status_code >= 500)
-        )
-        or (
-            isinstance(exception, requests.exceptions.HTTPError)
-            and (exception.response is None or exception.response.status_code >= 500)
-        )
-        or isinstance(exception, requests.exceptions.ConnectionError)
-        or isinstance(exception, requests.exceptions.Timeout)
-        or isinstance(exception, requests.exceptions.TooManyRedirects)
+    elif isinstance(exception, httpx.HTTPError) and (
+        exception.response is None or exception.response.status_code >= 500
     ):
         # NOTE(sileht): We already retry locally with urllib3, so if we get there, Github
         # is in a really bad shape...
