+++ setup.cfg
@@ -30,7 +30,6 @@ install_requires =
     celery[redis,gevent]
     setproctitle
     cryptography
-    requests
     aioredis
     redis!=3.4.1,!=3.4.0
     hiredis
@@ -53,6 +52,7 @@ test =
     pytest-cov
     pygithub>=1.43.8
     vcrpy>=1.12.0
+    requests
     pifpaf
 docs =
     sphinx
